requirements.txt
Flask
google-auth
google-oauth
ldap3